Move code and make x509 independent of data models
